requirements.txt
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.2
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
curl_cffi==0.13.0
distro==1.9.0
Django==6.0
django-allauth==65.13.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
feedparser==6.0.12
frozendict==2.4.7
google-auth==2.45.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multitasking==0.0.12
numpy==2.4.0
nvidia-nccl-cu12==2.29.2
openai==2.14.0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
razorpay==2.0.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlparse==0.5.5
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
websockets==15.0.1
wheel==0.45.1
xgboost==3.1.2
yfinance==1.0
